Fix circular deps pt 2
