+++ packages/web/src/components/nav/desktop/PlaylistLibrary/PlaylistFolderNavItem.tsx
@@ -23,7 +23,8 @@ import { useToggle } from 'react-use'
 import { make, useRecord } from 'common/store/analytics/actions'
 import { Draggable, Droppable } from 'components/dragndrop'
 import { setFolderId as setEditFolderModalFolderId } from 'store/application/ui/editFolderModal/slice'
-import { DragDropKind, selectDraggingKind } from 'store/dragndrop/slice'
+import { selectDraggingKind } from 'store/dragndrop/selectors'
+import { DragDropKind } from 'store/dragndrop/slice'
 import { useSelector } from 'utils/reducer'
 
 import { LeftNavLink } from '../LeftNavLink'
