+++ packages/web/src/pages/upload-page/UploadFormScrollContext.tsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react'
+
+export const UploadFormScrollContext = createContext(() => {})
