+++ packages/web/src/pages/upload-page/fields/MultiTrackSidebar.tsx
@@ -13,7 +13,7 @@ import DynamicImage from 'components/dynamic-image/DynamicImage'
 import layoutStyles from 'components/layout/layout.module.css'
 import { Text } from 'components/typography'
 
-import { UploadFormScrollContext } from '../UploadPage'
+import { UploadFormScrollContext } from '../UploadFormScrollContext'
 import { useIndexedField } from '../hooks'
 import { SingleTrackEditValues, TrackEditFormValues } from '../types'
 import { UploadPreviewContext } from '../utils/uploadPreviewContext'
