+++ packages/web/src/pages/upload-page/fields/availability/usdc-purchase-gated/UsdcPurchaseFields.tsx
@@ -17,7 +17,7 @@ import { TextField, TextFieldProps } from 'components/form-fields'
 import layoutStyles from 'components/layout/layout.module.css'
 import { Text } from 'components/typography'
 
-import { PREVIEW, PRICE } from '../../AccessAndSaleField'
+import { PREVIEW, PRICE } from '../../types'
 
 import styles from './UsdcPurchaseFields.module.css'
