+++ packages/web/src/pages/upload-page/forms/EditTrackForm.tsx
@@ -12,8 +12,8 @@ import IconCaretLeft from 'assets/img/iconCaretLeft.svg'
 import layoutStyles from 'components/layout/layout.module.css'
 import { NavigationPrompt } from 'components/navigation-prompt/NavigationPrompt'
 import { Text } from 'components/typography'
-import { UploadFormScrollContext } from 'pages/upload-page/UploadPage'
 
+import { UploadFormScrollContext } from '../UploadFormScrollContext'
 import { AnchoredSubmitRow } from '../components/AnchoredSubmitRow'
 import { PreviewButton } from '../components/PreviewButton'
 import { AccessAndSaleField } from '../fields/AccessAndSaleField'
